# improve performance of hashing seriesbytag with fnv