This PR addresses the poor performance of `BySeriesWithTagsFnv`. It moves the code from `fnv` into the method itself which reduces our allocations to 1 and speeds things up quite a bit.

It also adds the tests back into `schema`, and creates a new test specifically for the old and new methods used to obtain the partition with `fnv` in this PR.

These benchmarks were not performed in a suitable environment for benchmarking, but I think the difference is large enough to not warrant it.

## Benchcmp
```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkPartitionByOrg-12                   64.7          60.7          -6.18%
BenchmarkPartitionBySeries-12                60.0          56.9          -5.17%
BenchmarkPartitionBySeriesWithTags-12        629           630           +0.16%
BenchmarkPartitionBySeriesWithTagsFnv-12     1508          396           -73.74%

benchmark                                    old allocs     new allocs     delta
BenchmarkPartitionByOrg-12                   3              3              +0.00%
BenchmarkPartitionBySeries-12                2              2              +0.00%
BenchmarkPartitionBySeriesWithTags-12        2              2              +0.00%
BenchmarkPartitionBySeriesWithTagsFnv-12     23             1              -95.65%

benchmark                                    old bytes     new bytes     delta
BenchmarkPartitionByOrg-12                   16            16            +0.00%
BenchmarkPartitionBySeries-12                20            20            +0.00%
BenchmarkPartitionBySeriesWithTags-12        112           112           +0.00%
BenchmarkPartitionBySeriesWithTagsFnv-12     488           32            -93.44%
```

## Original
```
BenchmarkPartitionByOrg-12                      30000000                64.7 ns/op            16 B/op          3 allocs/op
BenchmarkPartitionBySeries-12                   20000000                60.0 ns/op            20 B/op          2 allocs/op
BenchmarkPartitionBySeriesWithTags-12            2000000               629 ns/op             112 B/op          2 allocs/op
BenchmarkPartitionBySeriesWithTagsFnv-12         1000000              1508 ns/op             488 B/op         23 allocs/op
```

## New
```
BenchmarkPartitionByOrg-12                      20000000                60.7 ns/op            16 B/op          3 allocs/op
BenchmarkPartitionBySeries-12                   30000000                56.9 ns/op            20 B/op          2 allocs/op
BenchmarkPartitionBySeriesWithTags-12            2000000               630 ns/op             112 B/op          2 allocs/op
BenchmarkPartitionBySeriesWithTagsFnv-12         5000000               396 ns/op              32 B/op          1 allocs/op
```